Remove --skip-refresh option from scripts/update-charts-dep.sh script
Helm is failing to download dependencies from external chart repos with this option specified, so we need to remove it.
One of such dependencies is reporting templates helm chart, which is included as a dependency for reporting service.
Removing this option shouldn't affect anything, as it is probably used for optimization purposes only.